Finished skeleton of communication with debug adapter.
This is merely the framework through which communication with the debug adapter over STDIO is begun.
Despite being the most popular JSON parsing library for Rust by far, serde-json is deliberately not
used here, because JSON, being a loosely typed language, doesn't translate to Rust very well, which
necessitates the definition of a Rust struct for every single message type that kak-dap is meant to
handle. While kak-lsp can do this with the lsp-types crate, no such crate exists for DAP, and I am
not defining every message type myself, which is where json-rust comes in. This crate doesn't attempt
to package JSON objects into neat Rust structs, but rather constructs generic objects that we can parse
with simple duck-typing techniques. Simply reading the "type" and "command" fields is enough to determine
the type of message we are dealing with, and therefore how they should be parsed.