Created basic "UI" setup and reverted to std::thread threads instead of crossbeam_util::thread threads.
It's about time I learned how ownership in Rust works. Using the threads
in the standard library turned out to be a lot easier than I thought - any
variables that need to be used by a thread are now "owned" by the thread,
and cannot be used anymore outside of the thread, so cloning variables that
we need to keep track of is nice. Also, the Arc datatype could help with
syncing across threads, should that prove necessary.

As for changes in functionality: there are now buffers that show up when
running dap-start (this is caused by dap-setup-ui), which will be where
the stack trace and variable views will go. The initial "telemetry"
messages from debugpy are (for now) shown in the debug buffer.