Added Context struct to ease passing around of "global" variables.
This is another idea lifted from kak-lsp - since we will want to pass around
many variables, such as the pipe to the debug adapter, the current session, and
the current request ID, it makes sense to place it in a struct that we can copy.
Heck, we're already shamelessly lifting the JSON-RPC parsing approach (sans serde)
from kak-lsp, so this is fine.