Added support for runInTerminal request. Debuggee now runs properly.
The runInTerminal request in the specification supports both an
integrated terminal and an external terminal. Since Kakoune doesn't
exactly have an integrated terminal, we simply ignore the 'kind'
argument in the runInTerminal request and spawn an external one
whether the debug adapter likes it or not ;). We also don't get
the pid of the spawned process, since the spec labels it as optional.

This also fixed the problems with the prior commit with the shell
suspending the Kakoune process. As it turns out, debugpy was trying
to run the debuggee in the background, and this caused some fighting
over the STDOUT and STDIN, causing the shell to suspend the process.