Added support for continue and "step" commands, and begun refactor.
Now, a hobbled form of "debugging" is possible. The user can begin
the program and step through it using the messages in the debug buffer
to determine what happened and when, but that's about it.

The next step would be to retrieve the stack frame upon receiving
the "stopped" event, so we can indicate where exactly execution
stopped. We could also use this information to obtain a variable
reference to pass to the "variables" request, so we can obtain
the variables in the current scope.