Added capability to view stack trace in stacktrace client upon receiving "stopped" event.
This implementation is fairly crude, simply printing the name of the function, the name
of the file and the line within the file for each entry in the stack trace. It does the
job, however, and was faily easy to implement.

Next on the docket is to implement the "variables" view - the main question for this would
be whether we are showing the variables within the current scope, or if we are showing variables
from the global scope.