First stab at displaying top-level variables in variables buffer.
This doesn't make any attempt at expanding variables with "structure",
such as arrays, structs and classes. Ideally, the user will be able to
control the expanding of these variables, clicking <Enter> on the line
containing the variable to expand. Due to simplicity, we won't support
retracting the variables in version 1.0.0. While primitive, this
approach should just about get the job done.