Begun complete rewrite of variables handling.
The way that the variables buffer is currently handled in kak-dap
is extremely ghetto and doesn't scale at all. For every variable
found by the "variables" request, a separate 'kak -p' call is made
to populate the buffer. This could potentially result in massive
slowdowns every time a "stopped" event was encountered. Also, it
is nearly impossible to implement "collapsing" of variables with
the current system.

The way we will combat this is to keep track of how many variables
requests are pending, and only print to the buffer once they've
all been serviced - that way we only need one 'kak -p' call and
the order in which the variables are added to the buffer will be
consistent.

This will also allow for a more robust implementation of expanding
and collapsing structured variables. Since we are storing the variables
and scopes found, we can check this whenever we service a "variables"
request, allowing for persisting of variable "collapsed" status.

In all, while this rewrite will add considerable complexity to the
way variables are handled in kak-dap, it will be far more maintainable
and scalable than the old implementation, and is therefore something
I consider worth doing before the 1.0.0 release.