Added list of pending requests to context struct.
OK, I need to stop procrastinating and actually get this done.

Since the 'variables' request's response doesn't maintain any information
about the request that spawned it apart from the sequence number, we need
another way to obtain information about where this 'variables' response came
from. Maintaining a list of requests we have sent to the adapter allows us
to do just that. This will allow us to search the list of pending requests
whenever we receive a 'variables' response to get the request we sent,
ultimately allowing us to construct a heirarchy of variables we can render
in the variables view.

The goal of all this is to allow us to construct a heirarchy of variables,
with scopes on the top level, and each scope/variable maintaining a list of
child variables. We can then use this heirarchy to send one 'kak -p' call
to the kakoune session to render the entire variable heirarchy all at once,
rather than piecemeal like before. This will also ultimately allow us to
keep certain variables expanded between 'stopped' events (eventually).