Complete redesign of how variables and scopes are represented and stored.
With the old approach of having each scope/variable store an array of sub-variables,
Rust's borrow checker was throwing errors regardless of which approach was taken
to retrieve the parent of any given variable. To solve this, we will simply define
an array of scopes and an array of variables, and thus any searches for parent
variables will be a simple linear search.