Display scopes in variables buffer.
Once we are finished processing all the pending "variables" requests, we
can begin rendering them in the variables buffer. To save on 'kak -p' calls,
we slowly build the entire command string to send to Kakoune and then send
it as a single command. This also allows us to ensure that the order in which
the scopes/variables are written to the buffer is consistent.

Currently, only the scopes are displayed in the variables buffer. However, this
should be easy to solve. We can use the variablesReference value of the scope
object to query the variables array for child variables, and render those below
the desired scope. It isn't the most efficient strategy in the world, of course,
but since I'm new to Rust, I needed something that worked with Rust's borrow checker.