Render '+' symbol on variables that can be expanded.
We are going to go with a fairly straightforward solution for expanding
variables. Essentially, every line in the "variables" buffer that contains
a variable will start with its variable reference. We will use this variable
reference, along with whether this variable is expanded or expanded, to
determine how to respond.

First, if the variable isn't expanded, we will send a single "variables"
request to the debug adapter to retrieve the child variables of the variable
we are expanding. This will then repopulate the variables buffer as before.

If the variable is expanded, however, we will simply remove all its child
variables from the variables list and re-populate the variables buffer.