Fixed buggy behavior in expanding variables containing "'" symbol.
This fix turned out to be a lot easier than expected - essentially, all
that needed to be done was to escape each instance of the "'" symbol by
doubling it up, so that Kakoune wouldn't interpret them as ends of strings.

With this, the variables support in kak-dap is feature-complete as far
as the initial release goes - still a bit clunky, but it works. Next on the
docket is having breakpoints sent from the editor rather than hardcoded in
the kak-dap binary.