Reverted session passing to happen via command line args.
On second thought, passing the Kakoune session via environment variables was a mistake.
While we can guarantee that certain command line arguments be passed to the application,
we can't guarantee that environment variables will be set. This necessitated ugly
fallbacks when the environment variable wasn't defined that would not work in 100%
of cases (if the session happened to be named "none", kak-dap would fail). Reverting to
using command line arguments to pass the session to the program allows us to check if the
necessary command line argument was defined, and if not, abort the application.

The biggest question now is how we're going to pass the currently set breakpoints to kak-dap
(for this initial release, we aren't going to allow changing the breakpoints at runtime).
The current best idea I have at the moment is to dump them into a file that kak-dap then
reads to retrieve the initial breakpoint data. This would be a temp file, since we would then
need to clean it up after the breakpoints are read. While this solution seems a tad clunky
and inelegant, I can't come up with a better approach at present.