Changed kak-dap communication method from stdin to sockets.
This is currently very rudimentary - it opens a socket at a fixed path,
regardless of your current Kakoune session or user, which means it likely
won't work in a multi-user environment. It also doesn't clean up the socket
upon closing the application. Finally, this creates a dependency on the user
having socat installed, which is by no means guaranteed.

Overall, while this does currently work, there is certainly room for improvement.