Remove socket file upon exiting kak-dap.
This still doesn't define separate socket files for each session,
but this should work for the time being. There's still a small flaw
with the current system: if kak-dap isn't exited through the "stop"
command, the socket file isn't erased, which will cause kak-dap to
crash upon next invocation. We may have to look at kak-lsp for ideas
on how to fix this.