Changed message format from "command args" to JSON; close issue #8
This change allows for more flexibility in the commands
we send to the kak-dap binary. Before, we needed to keep
everything on one line, which may have caused issues down the road.
Now, however, we are able to define messages with a flexible
JSON-based format that allows us to pass whatever arguments we need.