Fixed dap-stop and send breakpoints to kak-dap session.
Currently, breakpoints are sent by writing them to a file
and deleting the file once they're read into a JSON object.
The way this is implemented, however, will cause problems with
filepaths containing spaces. I unfortunately can't think of
a better way to do this, though, since I'm by no means
an expert in POSIX shell.

The next step for this would be to take the JSON value and
extract the "breakpoints" value, space separating them and
extract the line and filename information for use in
SetBreakpoint requests. After this is done, we can remove
the hardcoded breakpoint, as we can now set the breakpoints
ourselves, and issue #4 can be closed.