Handle "terminated" and "exited" events; teardown UI from Rust side
In addition to terminating the debug session when the debuggee
terminates, this fixes a longstanding bug - when the debug session
is terminated from a client other than the jump client, the location
flag is not reset. Wezterm kind of freaks out when this happens, but
I think that's due to the Wezterm integration plugin potentially being
a bit buggy.