Implemented search for debug configuration file.
At this point, we have the path to the debug configuration
file that will contain the debug adapter to launch and the
arguments to the "launch" request. Now, all we have to do is
read the file into a JSON object, and we can use that information
to start the desired debug adapter and set its launch args, and
issue #5 will be resolved.