Fixed cpp demo and fixed threadId not being correct in most instances
So, it turns out vscode-cpptools does work for c/c++ apps - we were just
doing a bunch of things wrong. For one thing, vscode-cpptools really
cares about the "type" parameter - it must be "cppdgb", or it won't work.

Additionally, we were just passing "1" as the threadId for various
requests, which is obviously not protocol compliant. While this was working
by accident with debugpy and the webfreak debugger, vscode-cpptools used
threadId values other than 1, which obviously broke it.

Since all of the above issues have been addressed, we are now compatible
with adapters for c, c++, Rust (in theory - not actually tested), and
Python. I need to update that wiki page.