Only send configurationDoneRequest if the capability is advertised by the adapter.
This should resolve scenarios where the reason the adapter doesn't work is because
the adapter doesn't support the configurationDoneRequest request. Bash-debug still
doesn't work with this, though, because the command it tries to run is invalid in POSIX
shell, which is what kak-dap uses to try to run the command. So for now, I'm writing
off support for that adapter as an impossibility.