Properly start debug adapter; bugfixes.
There were a LOT of bugs and architectural issues that needed to be
corrected here, from the adapter code not sending/receiving messages
properly to the Kakoune socket not working, etc. This overhaul fixes
all of these issues, and should make implementing proper debugging
significantly easier.

At this point, all that needs to happen before debugging can begin
is to send the runInTerminal response to the adapter, and then
the debuggee can begin. After that, it's mainly just implementing
the stack trace and variable handling, and implementing the various
commands from the Kakoune session. Still a ways to go, but we're
getting there.