Debuggee launches; stopped events trigger retrieving threads
At this point, we have begun debugging. After receiving threads,
the stack trace of the "current thread" needs to be retrieved,
and the variable heirarchy of that stack frame needs to be
retrieved as well. Once this is accomplished, we'll have the
basics working. Obviously there's still a fair bit to go at
that point, but we're getting there.