Added support for evaluating expressions.
At this point, we have pretty much achieved feature parity with the
Rust implementation. We still need to handle the "stopped" and
"terminated" events, though, so we exit properly when the debuggee
exits. Additionally, there are some extra features that still
need to be implemented, such as support for choosing which of the
defined debug configurations to use in any given debugging session,
as well as install scripts to install the debug adapters defined in
the adapters.yaml file shipped with kak-dap.