Added requirements file; added cpptools installer
This was actually a bit of an ordeal, due to how the
vscode-cpptools debug adapter works. Firstly, upon install the
adapter itself isn't executable, so we have to set those bits
ourselves (and so does the extension itself lol). Secondly, the
adapter is extremely picky about which value is passed to
adapterID, since it uses that to load a file it needs. Ugh.