Removed all dependencies from kak-dap.
As nice as the xdg and schema packages were, we don't actually
need them to implement the XDG specification and json validation,
respectively. As for yaml, it's more readable than json but is
a substantially more complex language, and is vulnerable to exploits
such as the million laughs attack. Json doesn't have these issues,
and is built into the Python stdlib.

With this, we are now dependency free, with cloning the plugin being
the only required step to begin using it.