Removed readapt adapter support; modified debugpy adapter
Unfortunately, the readapt adapter is no longer maintained,
and lacks a crucial feature for kak-dap: runInTerminal. Instead,
users can only view stdout through the debug buffer, and stdin
cannot be written to at all, which is a much worse experience.

Additionally, we want to limit how much the user's overall development
environment is impacted by our adapter installers, and as a result
using package managers to install readapt and debugpy was a no-go.
Debugpy supports the use case of downloading the source and running
it, fortunately.